Break up commands
