+++ src/index.ts
@@ -1,4 +1,8 @@
 export * from './Color';
+export * from './Documents/Commands';
+export * from './Documents/ConfigDocument';
+export * from './Documents/Document';
+export * from './Documents/LabelDocument';
 export * from './Documents/ReadyToPrintDocuments';
 export * from './Printers/Communication/LineBreakTransformer';
 export * from './Printers/Communication/PrinterCommunication';
